Fix bug on let-in expression not uncomputing the assignments due to exceptions beign launched
The assignments are computed, but when an error raises, they are not uncomputed.
Thus some variables may persist on the Environment when they should not, leading to
bugs.
This code evidentiates the bug:
let b = 2 in a;//Throws a syntactic error
b;//Prints 2 because b was never removed from the Environment.